+++ frontend/src/components/browser/search-combobox.vue
@@ -3,7 +3,7 @@
     ref="searchbox"
     v-model="query"
     v-model:menu="menu"
-    :items="queries"
+    :items="items"
     autofocus
     aria-label="search"
     clearable
@@ -27,6 +27,8 @@ import { mapActions, mapState } from "pinia";
 
 import { useBrowserStore } from "@/stores/browser";
 
+const MAX_ITEMS = 10;
+
 export default {
   name: "BrowserSearchCombobox",
   data() {
@@ -34,30 +36,36 @@ export default {
       mdiMagnify,
       menu: false,
       query: "",
+      items: [""],
     };
   },
   computed: {
     ...mapState(useBrowserStore, {
-      queries: (state) => state.page.queries,
       stateQ: (state) => state.settings.q,
     }),
   },
   watch: {
-    menu(to) {
-      if (!to) {
-        this.doSearch(this.query);
-      }
-    },
     stateQ(to) {
       this.query = to;
+      this.addToMenu(to);
     },
   },
   methods: {
     ...mapActions(useBrowserStore, ["setSettings"]),
+    addToMenu(q) {
+      if (!q || this.items.includes(q)) {
+        return;
+      }
+      const updateditems = this.items[0] === "" ? [] : this.items;
+      updateditems.unshift(q);
+      this.items = updateditems.slice(0, MAX_ITEMS);
+    },
     doSearch() {
+      console.log("doSearch()");
       this.menu = false;
       const q = this.query ? this.query.trim() : "";
       this.setSettings({ q });
+      this.addToMenu(q);
     },
   },
 };